scripts: do not restart kubelet on install/uninstall
On commit d5ef5600e03c it was added in the instruction to restart kubelet
on both post-install and post-uninstall. That seems to have caused the
post-uninstall daemon to get stuck on termination state. From what I
understood, once restarted, kubelet goes into a state where it cannot
umount the volumes of the post-uninstall daemon, hence the pod keeps
terminating forever.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>